flask-app/requirements.txt
boto3==1.14.16
botocore==1.17.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.13.0
repertorio==0.0.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
spotipy==2.12.0
urllib3==1.25.9
Werkzeug==1.0.1
